Move JsonWebToken to Helper
