--- AlkemyWallet/Core/Helper/ApiHelper.cs
@@ -1,6 +1,0 @@
-﻿namespace AlkemyWallet.Core.Helper
-{
-    public class ApiHelper
-    {
-    }
-}
